requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.3.0
click==8.1.3
coverage==6.4.1
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
traitlets==5.3.0
wcwidth==0.2.5
